requirements.txt
langchain-community
pypdf
langchain-huggingface
faiss-cpu
torch
tqdm
google-generativeai
typing
pydantic
tf-keras
sentence-transformers
fastapi
uvicorn
python-multipart
requests
python-docx
aiohttp
nest-asyncio
lxml
python-dotenv